- Bulk upload dialog uses the site sent as prop instead directly grabbing it from state on its own - Sites state initialization reads cookie to initialize the state with the selected site if there is one - Update bulk upload dialog copy - Published 1.0.0-alpha.1
