+++ ui/app/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@craftercms/studio-ui",
-  "version": "1.0.0-alpha",
+  "version": "1.0.0-alpha.1",
   "private": true,
   "repository": {
     "type": "git",
