+++ ui/app/src/state/reducers/env.ts
@@ -17,14 +17,21 @@
 import { createReducer } from '@reduxjs/toolkit';
 import { GlobalState } from '../../models/GlobalState';
 
-export const initialState: GlobalState['env'] = {
-  AUTHORING_BASE: `${window.location.origin}/studio`,
-  GUEST_BASE: window.location.origin,
+const initialState: GlobalState['env'] = ((origin: string) => ({
+  AUTHORING_BASE: `${origin}/studio`,
+  GUEST_BASE: origin,
   PREVIEW_LANDING_BASE: `/studio/preview-landing`,
   SITE_COOKIE: 'crafterSite',
   XSRF_CONFIG_ARGUMENT: '_csrf',
   XSRF_CONFIG_HEADER: 'X-XSRF-TOKEN'
-};
+}))(
+  process.env.NODE_ENV === 'production'
+    ? window.location.origin
+    : window.location.origin.replace(
+      process.env.REACT_APP_DEV_SERVER_PORT ?? '3000',
+      '8080'
+    )
+);
 
 const reducer = createReducer<GlobalState['env']>(initialState, {});
