+++ ui/app/src/translations/locales/en.json
@@ -52,7 +52,7 @@
   "browseCMIS.cmis": "CMIS",
   "bulkUpload.dropHere": "Drop files here or <span>browse</span>",
   "bulkUpload.filesProgression": "{start}/{end}",
-  "bulkUpload.subtitle": "Drop the desired files from your desktop into the browser window.",
+  "bulkUpload.subtitle": "Drop the desired files from your desktop into the space below.",
   "bulkUpload.title": "Bulk Upload",
   "bulkUpload.uploadInProgress": "Uploads are still in progress. Leaving this page would stop the pending uploads. Are you sure you wish to leave?",
   "bulkUploadConfirmDialogMessages.description": "There is still a bulk upload in progress. Only one bulk upload can be executed at the same time.",
