+++ ui/app/src/translations/locales/ko.json
@@ -52,7 +52,7 @@
   "browseCMIS.cmis": "CMIS",
   "bulkUpload.dropHere": "여기에 파일을 놓거나 <span>찾아보기</span>",
   "bulkUpload.filesProgression": "{start}/{end}",
-  "bulkUpload.subtitle": "데스크탑에서 브라우저 창으로 원하는 파일을 놓습니다.",
+  "bulkUpload.subtitle": "바탕 화면에서 원하는 공간을 아래 공간에 놓습니다.",
   "bulkUpload.title": "일괄 업로드",
   "bulkUpload.uploadInProgress": "업로드가 아직 진행 중입니다. 이 페이지를 떠나면 보류중인 업로드가 중지됩니다. 떠나시겠습니까?",
   "bulkUploadConfirmDialogMessages.description": "여전히 대량 업로드가 진행 중입니다. 한 번에 하나의 일괄 업로드 만 실행할 수 있습니다.",
